Replace ContinuousMLPPolicy
