+++ examples/tf/ddpg_pendulum.py
@@ -18,7 +18,7 @@ from garage.replay_buffer import SimpleReplayBuffer
 from garage.tf.algos import DDPG
 from garage.tf.envs import TfEnv
 from garage.tf.experiment import LocalTFRunner
-from garage.tf.policies import ContinuousMLPPolicyWithModel
+from garage.tf.policies import ContinuousMLPPolicy
 from garage.tf.q_functions import ContinuousMLPQFunction
 
 
@@ -29,10 +29,10 @@ def run_task(snapshot_config, *_):
 
         action_noise = OUStrategy(env.spec, sigma=0.2)
 
-        policy = ContinuousMLPPolicyWithModel(env_spec=env.spec,
-                                              hidden_sizes=[64, 64],
-                                              hidden_nonlinearity=tf.nn.relu,
-                                              output_nonlinearity=tf.nn.tanh)
+        policy = ContinuousMLPPolicy(env_spec=env.spec,
+                                     hidden_sizes=[64, 64],
+                                     hidden_nonlinearity=tf.nn.relu,
+                                     output_nonlinearity=tf.nn.tanh)
 
         qf = ContinuousMLPQFunction(env_spec=env.spec,
                                     hidden_sizes=[64, 64],
