+++ src/garage/tf/policies/__init__.py
@@ -14,8 +14,6 @@ from garage.tf.policies.categorical_mlp_policy import CategoricalMLPPolicy
 from garage.tf.policies.categorical_mlp_policy_with_model import (
     CategoricalMLPPolicyWithModel)
 from garage.tf.policies.continuous_mlp_policy import ContinuousMLPPolicy
-from garage.tf.policies.continuous_mlp_policy_with_model import (
-    ContinuousMLPPolicyWithModel)
 from garage.tf.policies.deterministic_mlp_policy import DeterministicMLPPolicy
 from garage.tf.policies.discrete_qf_derived_policy import (
     DiscreteQfDerivedPolicy)
@@ -31,7 +29,6 @@ __all__ = [
     'CategoricalGRUPolicyWithModel', 'CategoricalLSTMPolicy',
     'CategoricalLSTMPolicyWithModel', 'CategoricalMLPPolicy',
     'CategoricalMLPPolicyWithModel', 'ContinuousMLPPolicy',
-    'ContinuousMLPPolicyWithModel', 'DiscreteQfDerivedPolicy',
-    'DeterministicMLPPolicy', 'GaussianGRUPolicy', 'GaussianLSTMPolicy',
-    'GaussianGRUPolicyWithModel', 'GaussianMLPPolicy'
+    'DiscreteQfDerivedPolicy', 'DeterministicMLPPolicy', 'GaussianGRUPolicy',
+    'GaussianLSTMPolicy', 'GaussianGRUPolicyWithModel', 'GaussianMLPPolicy'
 ]
